+++ CHANGES.rst
@@ -5,13 +5,16 @@ Changelog
 1.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- We can define a dexterity adapter for just one field using field name as
+  adapter name.
+  [thomasdesvenain]
 
 
 1.1 (2014-06-19)
 ----------------
 
 - Rename search policy excelexport.search to avoid conflict with 'search' view.
+  [thomasdesvenain]
 
 
 1.0 (2014-06-02)